+++ src/Components/Forms/Base/FormComponentBase.cs
@@ -1,25 +1,37 @@
 ﻿using Microsoft.AspNetCore.Components;
 using Microsoft.AspNetCore.Components.Forms;
 using System.Diagnostics.CodeAnalysis;
+using System.Linq.Expressions;
 
 namespace Tavenem.Blazor.Framework.Components.Forms;
 
 /// <summary>
 /// A base class for form components.
 /// </summary>
-public abstract class FormComponentBase<TValue> : InputBase<TValue>, IFormComponent
+public abstract class FormComponentBase<TValue> : ComponentBase, IDisposable, IFormComponent
 {
     private class DummyClass<T>
     {
         public T? This_field { get; set; }
     }
 
+    private readonly EventHandler<ValidationStateChangedEventArgs> _validationStateChangedHandler;
     private readonly AsyncAdjustableTimer _timer;
 
     private List<string>? _customValidationMessages;
+    private bool _hasInitializedParameters;
     private bool _disposedValue;
     private DummyClass<TValue>? _dummyModel;
     private bool _initialParametersSet;
+    private Type? _nullableUnderlyingType;
+    private bool _previousParsingAttemptFailed;
+    private ValidationMessageStore? _parsingValidationMessages;
+
+    /// <summary>
+    /// Custom HTML attributes for the component.
+    /// </summary>
+    [Parameter(CaptureUnmatchedValues = true)]
+    public IReadOnlyDictionary<string, object>? AdditionalAttributes { get; set; }
 
     /// <summary>
     /// Custom CSS class(es) for the component.
@@ -38,7 +50,7 @@ public abstract class FormComponentBase<TValue> : InputBase<TValue>, IFormCompon
 
     /// <summary>
     /// <para>
-    /// The validation message displayed when this field's <see cref="InputBase{TValue}.Value"/>
+    /// The validation message displayed when this field's <see cref="Value"/>
     /// cannot be converted to or from its string representation.
     /// </para>
     /// <para>
@@ -46,6 +58,17 @@ public abstract class FormComponentBase<TValue> : InputBase<TValue>, IFormCompon
     /// </para>
     /// </summary>
     [Parameter] public virtual string ConversionValidationMessage { get; set; } = "{0} could not be converted";
+
+    /// <summary>
+    /// <para>
+    /// Gets or sets the display name for this field.
+    /// </para>
+    /// <para>
+    /// This value is used when generating error messages when the input value fails to parse
+    /// correctly.
+    /// </para>
+    /// </summary>
+    [Parameter] public string? DisplayName { get; set; }
 
     /// <summary>
     /// Whether the current value of the HTML input failed to convert to or from the bound data
@@ -123,10 +146,28 @@ public abstract class FormComponentBase<TValue> : InputBase<TValue>, IFormCompon
     [Parameter] public Func<TValue?, object?, IAsyncEnumerable<string>>? Validation { get; set; }
 
     /// <summary>
-    /// The final value assigned to the class attribute, including component values and anything
-    /// assigned by the user in <see cref="InputBase{TValue}.AdditionalAttributes"/>.
+    /// Gets or sets the value of the input. This should be used with two-way binding.
     /// </summary>
-    protected virtual new string? CssClass => new CssBuilder(Class)
+    /// <example>
+    /// @bind-Value="model.PropertyName"
+    /// </example>
+    [Parameter] public TValue? Value { get; set; }
+
+    /// <summary>
+    /// Gets or sets a callback that updates the bound value.
+    /// </summary>
+    [Parameter] public EventCallback<TValue> ValueChanged { get; set; }
+
+    /// <summary>
+    /// Gets or sets an expression that identifies the bound value.
+    /// </summary>
+    [Parameter] public Expression<Func<TValue>>? ValueExpression { get; set; }
+
+    /// <summary>
+    /// The final value assigned to the class attribute, including component values and anything
+    /// assigned by the user in <see cref="AdditionalAttributes"/>.
+    /// </summary>
+    protected virtual string? CssClass => new CssBuilder(Class)
         .AddClassFromDictionary(AdditionalAttributes)
         .Add("form-field")
         .Add("modified", IsTouched)
@@ -136,11 +177,93 @@ public abstract class FormComponentBase<TValue> : InputBase<TValue>, IFormCompon
 
     /// <summary>
     /// The final value assigned to the style attribute, including component values and anything
-    /// assigned by the user in <see cref="InputBase{TValue}.AdditionalAttributes"/>.
+    /// assigned by the user in <see cref="AdditionalAttributes"/>.
     /// </summary>
     protected virtual string? CssStyle => new CssBuilder(Style)
         .AddStyleFromDictionary(AdditionalAttributes)
         .ToString();
+
+    /// <summary>
+    /// Gets or sets the current value of the input.
+    /// </summary>
+    protected TValue? CurrentValue
+    {
+        get => Value;
+        set
+        {
+            var hasChanged = !EqualityComparer<TValue>.Default.Equals(value, Value);
+            if (hasChanged)
+            {
+                Value = value;
+                _ = ValueChanged.InvokeAsync(Value);
+                EditContext?.NotifyFieldChanged(FieldIdentifier);
+            }
+        }
+    }
+
+    /// <summary>
+    /// Gets or sets the current value of the input, represented as a string.
+    /// </summary>
+    protected string? CurrentValueAsString
+    {
+        get => FormatValueAsString(CurrentValue);
+        set
+        {
+            _parsingValidationMessages?.Clear();
+
+            bool parsingFailed;
+
+            if (_nullableUnderlyingType != null && string.IsNullOrEmpty(value))
+            {
+                // Assume if it's a nullable type, null/empty inputs should correspond to default(T)
+                // Then all subclasses get nullable support almost automatically (they just have to
+                // not reject Nullable<T> based on the type itself).
+                parsingFailed = false;
+                CurrentValue = default!;
+            }
+            else if (TryParseValueFromString(value, out var parsedValue, out var validationErrorMessage))
+            {
+                parsingFailed = false;
+                CurrentValue = parsedValue!;
+            }
+            else
+            {
+                parsingFailed = true;
+
+                // EditContext may be null if the input is not a child component of EditForm.
+                if (EditContext is not null)
+                {
+                    _parsingValidationMessages ??= new ValidationMessageStore(EditContext);
+                    _parsingValidationMessages.Add(FieldIdentifier, validationErrorMessage);
+
+                    // Since we're not writing to CurrentValue, we'll need to notify about modification from here
+                    EditContext.NotifyFieldChanged(FieldIdentifier);
+                }
+            }
+
+            // We can skip the validation notification if we were previously valid and still are
+            if (parsingFailed || _previousParsingAttemptFailed)
+            {
+                EditContext?.NotifyValidationStateChanged();
+                _previousParsingAttemptFailed = parsingFailed;
+            }
+        }
+    }
+
+    [CascadingParameter] private EditContext? CascadedEditContext { get; set; }
+
+    /// <summary>
+    /// Gets the associated <see cref="Microsoft.AspNetCore.Components.Forms.EditContext"/>. This
+    /// property is uninitialized if the input does not have a parent <see cref="Framework.Form"/>
+    /// or <see cref="EditForm"/>.
+    /// </summary>
+    protected EditContext EditContext { get; set; } = default!;
+
+    /// <summary>
+    /// Gets the <see cref="Microsoft.AspNetCore.Components.Forms.FieldIdentifier"/> for the bound
+    /// value.
+    /// </summary>
+    protected internal FieldIdentifier FieldIdentifier { get; set; }
 
     /// <summary>
     /// <para>
@@ -157,7 +280,11 @@ public abstract class FormComponentBase<TValue> : InputBase<TValue>, IFormCompon
     /// <summary>
     /// Constructs a new instance of <see cref="FormComponentBase{TValue}"/>.
     /// </summary>
-    protected FormComponentBase() => _timer = new(OnTimerAsync, 300);
+    protected FormComponentBase()
+    {
+        _timer = new(OnTimerAsync, 300);
+        _validationStateChangedHandler = OnValidateStateChanged;
+    }
 
     /// <inheritdoc/>
     public override async Task SetParametersAsync(ParameterView parameters)
@@ -168,15 +295,46 @@ public abstract class FormComponentBase<TValue> : InputBase<TValue>, IFormCompon
             ValueExpression = () => _dummyModel.This_field!;
         }
 
-        await base.SetParametersAsync(parameters);
+        parameters.SetParameterProperties(this);
+
+        if (!_hasInitializedParameters)
+        {
+            // This is the first run
+            // Could put this logic in OnInit, but its nice to avoid forcing people who override OnInit to call base.OnInit()
+
+            if (ValueExpression is null)
+            {
+                throw new InvalidOperationException($"{GetType()} requires a value for the 'ValueExpression' parameter. Normally this is provided automatically when using 'bind-Value'.");
+            }
+
+            FieldIdentifier = FieldIdentifier.Create(ValueExpression);
+
+            if (CascadedEditContext is not null)
+            {
+                EditContext = CascadedEditContext;
+                EditContext.OnValidationStateChanged += _validationStateChangedHandler;
+            }
+
+            _nullableUnderlyingType = Nullable.GetUnderlyingType(typeof(TValue));
+            _hasInitializedParameters = true;
+        }
+        else if (CascadedEditContext != EditContext)
+        {
+            // Not the first run
+
+            // We don't support changing EditContext because it's messy to be clearing up state and event
+            // handlers for the previous one, and there's no strong use case. If a strong use case
+            // emerges, we can consider changing this.
+            throw new InvalidOperationException($"{GetType()} does not support changing the {nameof(EditContext)} dynamically.");
+        }
+
+        UpdateAdditionalValidationAttributes();
+
+        await base.SetParametersAsync(ParameterView.Empty);
 
         if (!_initialParametersSet)
         {
             InitialValue = Value;
-            if (EditContext is not null)
-            {
-                EditContext.OnValidationStateChanged += OnValidationStateChanged;
-            }
             _initialParametersSet = true;
         }
     }
@@ -188,6 +346,19 @@ public abstract class FormComponentBase<TValue> : InputBase<TValue>, IFormCompon
         {
             Form?.Add(this);
         }
+    }
+
+    /// <inheritdoc/>
+    void IDisposable.Dispose()
+    {
+        // When initialization in the SetParametersAsync method fails, the EditContext property can remain equal to null
+        if (EditContext is not null)
+        {
+            EditContext.OnValidationStateChanged -= _validationStateChangedHandler;
+        }
+
+        Dispose(disposing: true);
+        GC.SuppressFinalize(this);
     }
 
     /// <summary>
@@ -271,7 +442,7 @@ public abstract class FormComponentBase<TValue> : InputBase<TValue>, IFormCompon
     /// Performs application-defined tasks associated with freeing, releasing, or resetting
     /// unmanaged resources.
     /// </summary>
-    protected override void Dispose(bool disposing)
+    protected virtual void Dispose(bool disposing)
     {
         if (!_disposedValue)
         {
@@ -279,18 +450,30 @@ public abstract class FormComponentBase<TValue> : InputBase<TValue>, IFormCompon
             {
                 _timer.Dispose();
                 Form?.Remove(this);
-                if (EditContext is not null)
-                {
-                    EditContext.OnValidationStateChanged -= OnValidationStateChanged;
-                }
             }
 
             _disposedValue = true;
         }
     }
 
-    /// <inheritdoc/>
-    protected override bool TryParseValueFromString(
+    /// <summary>
+    /// Formats the value as a string. Derived classes can override this to determine the formating
+    /// used for <see cref="CurrentValueAsString"/>.
+    /// </summary>
+    /// <param name="value">The value to format.</param>
+    /// <returns>A string representation of the value.</returns>
+    protected virtual string? FormatValueAsString(TValue? value) => value?.ToString();
+
+    /// <summary>
+    /// Parses a string to create an instance of <typeparamref name="TValue"/>. Derived classes can
+    /// override this to change how <see cref="CurrentValueAsString"/> interprets incoming values.
+    /// </summary>
+    /// <param name="value">The string value to be parsed.</param>
+    /// <param name="result">An instance of <typeparamref name="TValue"/>.</param>
+    /// <param name="validationErrorMessage">If the value could not be parsed, provides a validation
+    /// error message.</param>
+    /// <returns>True if the value could be parsed; otherwise false.</returns>
+    protected virtual bool TryParseValueFromString(
         string? value,
         [MaybeNullWhen(false)] out TValue result,
         [NotNullWhen(false)] out string? validationErrorMessage)
@@ -329,6 +512,37 @@ public abstract class FormComponentBase<TValue> : InputBase<TValue>, IFormCompon
         return success;
     }
 
+    /// <summary>
+    /// Returns a dictionary with the same values as the specified <paramref name="source"/>.
+    /// </summary>
+    /// <returns>
+    /// <see langword="true"/>, if a new dictionary with copied values was created; otherwise <see
+    /// langword="false"/>.
+    /// </returns>
+    private static bool ConvertToDictionary(IReadOnlyDictionary<string, object>? source, out Dictionary<string, object> result)
+    {
+        var newDictionaryCreated = true;
+        if (source is null)
+        {
+            result = new Dictionary<string, object>();
+        }
+        else if (source is Dictionary<string, object> currentDictionary)
+        {
+            result = currentDictionary;
+            newDictionaryCreated = false;
+        }
+        else
+        {
+            result = new Dictionary<string, object>();
+            foreach (var item in source)
+            {
+                result.Add(item.Key, item.Value);
+            }
+        }
+
+        return newDictionaryCreated;
+    }
+
     private protected void EvaluateDebounced()
     {
         if (!_disposedValue)
@@ -346,11 +560,62 @@ public abstract class FormComponentBase<TValue> : InputBase<TValue>, IFormCompon
 
     private Task OnTimerAsync() => InvokeAsync(ValidateAutoAsync);
 
-    private void OnValidationStateChanged(object? sender, ValidationStateChangedEventArgs e)
+    private void OnValidateStateChanged(object? sender, ValidationStateChangedEventArgs e)
     {
+        UpdateAdditionalValidationAttributes();
+
         if (!IsNested && EditContext is not null)
         {
             EvaluateDebounced();
+        }
+
+        StateHasChanged();
+    }
+
+    private void UpdateAdditionalValidationAttributes()
+    {
+        if (EditContext is null)
+        {
+            return;
+        }
+
+        var hasAriaInvalidAttribute = AdditionalAttributes?.ContainsKey("aria-invalid") == true;
+        if (EditContext.GetValidationMessages(FieldIdentifier).Any())
+        {
+            if (hasAriaInvalidAttribute)
+            {
+                // Do not overwrite the attribute value
+                return;
+            }
+
+            if (ConvertToDictionary(AdditionalAttributes, out var additionalAttributes))
+            {
+                AdditionalAttributes = additionalAttributes;
+            }
+
+            // To make the `Input` components accessible by default
+            // we will automatically render the `aria-invalid` attribute when the validation fails
+            // value must be "true" see https://www.w3.org/TR/wai-aria-1.1/#aria-invalid
+            additionalAttributes["aria-invalid"] = "true";
+        }
+        else if (hasAriaInvalidAttribute)
+        {
+            // No validation errors. Need to remove `aria-invalid` if it was rendered already
+
+            if (AdditionalAttributes!.Count == 1)
+            {
+                // Only aria-invalid argument is present which we don't need any more
+                AdditionalAttributes = null;
+            }
+            else
+            {
+                if (ConvertToDictionary(AdditionalAttributes, out var additionalAttributes))
+                {
+                    AdditionalAttributes = additionalAttributes;
+                }
+
+                additionalAttributes.Remove("aria-invalid");
+            }
         }
     }