+++ src/Components/Forms/Slider/Slider.razor.cs
@@ -1,5 +1,4 @@
 using Microsoft.AspNetCore.Components;
-using Microsoft.AspNetCore.Components.Forms;
 using System.Diagnostics.CodeAnalysis;
 using System.Globalization;
 using Tavenem.Blazor.Framework.Components.Forms;
@@ -32,8 +31,9 @@ public partial class Slider<TValue>
 
     /// <summary>
     /// <para>
-    /// The validation message displayed when this field's <see cref="InputBase{TValue}.Value"/>
-    /// cannot be converted to or from its string representation.
+    /// The validation message displayed when this field's <see
+    /// cref="FormComponentBase{TValue}.Value"/> cannot be converted to or from its string
+    /// representation.
     /// </para>
     /// <para>
     /// Default is "{DisplayName} must be a number".
@@ -128,15 +128,16 @@ public partial class Slider<TValue>
     [Parameter] public ThemeColor ThemeColor { get; set; }
 
     /// <summary>
-    /// Whether the bound <see cref="InputBase{TValue}.Value"/> should update whenever the value
-    /// changes (rather than on blur or when the enter key is pressed).
+    /// Whether the bound <see cref="FormComponentBase{TValue}.Value"/> should update whenever the
+    /// value changes (rather than on blur or when the enter key is pressed).
     /// </summary>
     [Parameter] public bool UpdateOnInput { get; set; }
 
     /// <summary>
     /// <para>
     /// When <see cref="UpdateOnInput"/> is true, this can be set to a number of milliseconds that
-    /// the component will wait before updating the bound <see cref="InputBase{TValue}.Value"/>.
+    /// the component will wait before updating the bound <see
+    /// cref="FormComponentBase{TValue}.Value"/>.
     /// </para>
     /// <para>
     /// Even when a debounce is used, the component will always update immediately on blur and when